Final commit

+++ pIII/FlightApp/app/src/main/java/backend/Admin.java
@@ -36,7 +36,7 @@ public class Admin extends User implements Serializable {
      * @param password  the password for this admin.
      */
     public Admin(String lastName, String firstName, String email,
-                 String address, int creditCardNumber, Date expiryDate,
+                 String address, long creditCardNumber, Date expiryDate,
                  String password) {
 
         super(lastName, firstName, email, address, creditCardNumber,
@@ -59,7 +59,7 @@ public class Admin extends User implements Serializable {
      * 															expiry date.
      */
     public Admin(String lastName, String firstName, String email,
-                 String address, int creditCardNumber, Date expiryDate) {
+                 String address, long creditCardNumber, Date expiryDate) {
         super(lastName, firstName, email, address, creditCardNumber,
                 expiryDate);
     }
